Fix byte cast for wall data. Closes #1252

+++ src/TEdit/ViewModel/WorldViewModel.Editor.cs
@@ -522,7 +522,7 @@ namespace TEdit.ViewModel
             }
 
             if (wall != null)
-                curTile.Wall = (byte)wall;
+                curTile.Wall = (ushort)wall;
 
             if (liquid != null)
             {
